Table top controllers on/off button fix.
The Table top controllers have 'All On' and 'All Off' buttons in
addition the the normal scene buttons.  For all three sizes of
controllers, the All On an All Off buttons are defined as
buttons 16 and 17.

The Node definitions for the 5 button and 10 button controllers
were using 6 & 7 and 11 & 12 for the All On and All Off buttons
so they didn't work.  Changed these two nodes to use the
correct button numbers 16 and 17.